requirements.txt
torch>=1.9.0
transformers>=4.12.0
datasets>=1.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
wandb>=0.12.0
tensorboard>=2.7.0
jupyter>=1.0.0
ipywidgets>=7.6.0
nltk>=3.6.0
spacy>=3.4.0
plotly>=5.3.0
PyYAML>=6.0 